requirements.txt
pandas
pymongo
python-dotenv
jupyter
matplotlib
seaborn
openpyxl

